Move artifacts.py out of API directory
